2025/adk-agentengine-basic/pyproject.toml
[project]
name = "adk-agentengine-basic"
version = "0.1.0"
description = "Hands-on asset for Agent Development Kit and Agent Engine basic level"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-adk>=1.8.0",
    "lxml>=6.0.0",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
]
